rename server to backend
