+++ docs/source/guides/getting-started/_examples/run_fastapi.py
@@ -1,7 +1,7 @@
 # :lines: 11-
 
 from idom import run
-from idom.server import fastapi as fastapi_server
+from idom.backend import fastapi as fastapi_server
 
 
 # the run() function is the entry point for examples
@@ -11,7 +11,7 @@ fastapi_server.configure = lambda _, cmpt: run(cmpt)
 from fastapi import FastAPI
 
 from idom import component, html
-from idom.server.fastapi import configure
+from idom.backend.fastapi import configure
 
 
 @component
