+++ docs/source/guides/getting-started/_examples/run_flask.py
@@ -1,7 +1,7 @@
 # :lines: 11-
 
 from idom import run
-from idom.server import flask as flask_server
+from idom.backend import flask as flask_server
 
 
 # the run() function is the entry point for examples
@@ -11,7 +11,7 @@ flask_server.configure = lambda _, cmpt: run(cmpt)
 from flask import Flask
 
 from idom import component, html
-from idom.server.flask import configure
+from idom.backend.flask import configure
 
 
 @component
