+++ docs/source/guides/getting-started/_examples/run_sanic.py
@@ -1,7 +1,7 @@
 # :lines: 11-
 
 from idom import run
-from idom.server import sanic as sanic_server
+from idom.backend import sanic as sanic_server
 
 
 # the run() function is the entry point for examples
@@ -11,7 +11,7 @@ sanic_server.configure = lambda _, cmpt: run(cmpt)
 from sanic import Sanic
 
 from idom import component, html
-from idom.server.sanic import configure
+from idom.backend.sanic import configure
 
 
 @component
