+++ src/idom/backend/flask.py
@@ -25,11 +25,11 @@ from simple_websocket import Server as WebSocket
 from werkzeug.serving import BaseWSGIServer, make_server
 
 import idom
+from idom.backend.types import Location
 from idom.core.hooks import Context, create_context, use_context
 from idom.core.layout import LayoutEvent, LayoutUpdate
 from idom.core.serve import serve_json_patch
 from idom.core.types import ComponentType, RootComponentConstructor
-from idom.server.types import Location
 from idom.utils import Ref
 
 from .utils import safe_client_build_dir_path, safe_web_modules_dir_path
