+++ tests/test_server/test_common.py
@@ -4,9 +4,9 @@ import pytest
 
 import idom
 from idom import html
-from idom.server import default as default_implementation
-from idom.server.types import Location
-from idom.server.utils import all_implementations
+from idom.backend import default as default_implementation
+from idom.backend.types import Location
+from idom.backend.utils import all_implementations
 from idom.testing import DisplayFixture, ServerFixture, poll
